Remove build system
