--- build-system/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "build-system"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-reqwest = { version = "0.12.12", features = ["blocking"] }
-flate2 = "1.0"
-tar = "0.4"
-indicatif = "0.17.9"
-serde = { version = "1.0.217", features = ["derive"] }
-serde_json = "1.0.134"
